# microservicio-catalogo-main/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Gestión de Catálogo","description":"Endpoints para administrar y consultar el catálogo de la biblioteca"}],"paths":{"/libros/{id}/disponibilidad":{"put":{"tags":["Gestión de Catálogo"],"summary":"Modificar el estado de disponibilidad de un libro","description":"Actualiza si un libro está disponible o no. Exclusivo para administradores del catálogo.","operationId":"actualizarDisponibilidad","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"boolean"}}},"required":true},"responses":{"200":{"description":"Estado del libro modificado exitosamente"},"400":{"description":"Error en los datos enviados para la actualización"},"403":{"description":"Acceso denegado por falta de autorización"}}}},"/libros/{id}":{"get":{"tags":["Gestión de Catálogo"],"summary":"Consultar información de un libro por su ID","description":"Devuelve los detalles de un libro específico. Requiere permisos de administrador o visualizador.","operationId":"obtenerLibro","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"Acceso restringido por falta de permisos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Libro"}}}},"404":{"description":"No se encontró un libro con ese ID","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Libro"}}}},"200":{"description":"Detalles del libro recuperados con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Libro"}}}}}}},"/libros/{id}/disponible":{"get":{"tags":["Gestión de Catálogo"],"summary":"Comprobar si un libro está disponible","description":"Indica si un libro puede ser prestado actualmente. Acceso para administradores y visualizadores.","operationId":"isLibroDisponible","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"Permisos insuficientes para realizar la consulta","content":{"*/*":{"schema":{"type":"boolean"}}}},"404":{"description":"El libro solicitado no existe en el catálogo","content":{"*/*":{"schema":{"type":"boolean"}}}},"200":{"description":"Estado de disponibilidad devuelto correctamente","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/libros/buscar":{"get":{"tags":["Gestión de Catálogo"],"summary":"Realizar una búsqueda de libros en el catálogo","description":"Permite encontrar libros según un criterio de búsqueda. Disponible para administradores y visualizadores.","operationId":"buscarLibros","parameters":[{"name":"criterio","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Búsqueda completada, lista de libros devuelta","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Libro"}}}}},"403":{"description":"Solicitud rechazada por falta de permisos","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Libro"}}}}}}}}},"components":{"schemas":{"Autor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"Categoria":{"type":"object","properties":{"nombre":{"type":"string"}}},"ISBN":{"type":"object","properties":{"isbn_value":{"type":"string"}}},"Libro":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/LibroId"},"titulo":{"type":"string"},"isbn":{"$ref":"#/components/schemas/ISBN"},"categoria":{"$ref":"#/components/schemas/Categoria"},"disponible":{"type":"boolean"},"autores":{"type":"array","items":{"$ref":"#/components/schemas/Autor"}}}},"LibroId":{"type":"object","properties":{"libroid_value":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}